修改autoId的类型

+++ seeds-game-model/src/main/java/com/seeds/game/dto/request/NftDistributeReq.java
@@ -12,7 +12,7 @@ public class NftDistributeReq {
     /**
      * nft全局id
      */
-    private Integer autoId;
+    private Long autoId;
     /**
      * 游戏服务角色id（serverRoleId）
      */
